+++ apps/eternal-e2e/src/pom/activation.pom.ts
@@ -1,9 +1,0 @@
-class Activation {
-  activate(code: string) {
-    cy.get('app-activate input').type(code);
-    cy.get('app-activate button[type=submit]').click();
-    cy.testid('activated');
-  }
-}
-
-export const activation = new Activation();