+++ apps/eternal/src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { UserLoaderGuard } from './core/user-loader.guard';
 import { HomeComponent } from './home.component';
 import { NewsletterComponent } from './newsletter/newsletter.component';
 
-export const APP_ROUTES: Routes = [
+export const appRoutes: Routes = [
   {
     path: '',
     canActivate: [UserLoaderGuard],
@@ -15,20 +15,16 @@ export const APP_ROUTES: Routes = [
       { path: 'home', redirectTo: '' },
       { path: 'newsletter', component: NewsletterComponent },
       {
-        path: 'security',
-        loadChildren: () => import('./security/security.module').then((m) => m.SecurityModule)
-      },
-      {
         path: 'customer',
-        loadChildren: () => import('./customer/customer.module').then((m) => m.CustomerModule)
+        loadChildren: () => import('./customer/customer.routes').then((m) => m.customerRoutes)
       },
       {
         path: 'holidays',
-        loadChildren: () => import('./holidays/holidays.module').then((m) => m.HolidaysModule)
+        loadChildren: () => import('./holidays/holidays.routes').then((m) => m.holidayRoutes)
       },
       {
         path: 'diary',
-        loadChildren: () => import('./diary/diary.routes.module').then((m) => m.DiaryRoutesModule)
+        loadChildren: () => import('./diary/diary.routes.module').then((m) => m.diaryRoutes)
       }
     ]
   }