+++ apps/eternal/src/app/core/loading.interceptor.ts
@@ -5,22 +5,22 @@ import {
   HttpRequest,
   HttpResponse
 } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 import { LoadingService } from '../shared/loading.service';
 
 @Injectable()
 export class LoadingInterceptor implements HttpInterceptor {
-  constructor(private loadingService: LoadingService) {}
+  #loadingService = inject(LoadingService);
 
   intercept(req: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
-    this.loadingService.loading();
+    this.#loadingService.loading();
 
     return next.handle(req).pipe(
       tap((event) => {
         if (event instanceof HttpResponse) {
-          this.loadingService.loaded();
+          this.#loadingService.loaded();
         }
       })
     );