+++ apps/eternal/src/app/diary/diaries/+state/diary.effects.ts
@@ -1,25 +1,28 @@
-import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { map, switchMap } from 'rxjs/operators';
 import { diaryActions } from './diary.actions';
 import { DiaryResponse, DiaryWithEntries } from './diary.reducer';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable()
 export class DiaryEffects {
+  #actions$ = inject(Actions);
+  #httpClient = inject(HttpClient);
+
   load$ = createEffect(() =>
-    this.actions$.pipe(
+    this.#actions$.pipe(
       ofType(diaryActions.load),
-      switchMap(() => this.httpClient.get<DiaryResponse>('/diary')),
+      switchMap(() => this.#httpClient.get<DiaryResponse>('/diary')),
       map((diaryResponse) => diaryActions.loadSuccess({ diaryResponse }))
     )
   );
 
   add$ = createEffect(() =>
-    this.actions$.pipe(
+    this.#actions$.pipe(
       ofType(diaryActions.add),
       switchMap(({ title, description }) =>
-        this.httpClient.post<DiaryWithEntries>('/diary', {
+        this.#httpClient.post<DiaryWithEntries>('/diary', {
           title,
           description
         })
@@ -29,10 +32,10 @@ export class DiaryEffects {
   );
 
   addEntry$ = createEffect(() =>
-    this.actions$.pipe(
+    this.#actions$.pipe(
       ofType(diaryActions.addEntry),
       switchMap(({ diaryId, content }) =>
-        this.httpClient.post('/diary', {
+        this.#httpClient.post('/diary', {
           diaryId,
           content
         })
@@ -40,6 +43,4 @@ export class DiaryEffects {
       map(() => diaryActions.addEntrySuccess())
     )
   );
-
-  constructor(private actions$: Actions, private httpClient: HttpClient) {}
 }