+++ apps/eternal/src/app/diary/diary.routes.module.ts
@@ -1,22 +1,20 @@
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
+import { importProvidersFrom } from '@angular/core';
+import { Routes } from '@angular/router';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule } from '@ngrx/store';
 import { DiaryEffects } from './diaries/+state/diary.effects';
 import { diaryFeatureKey, diaryReducer } from './diaries/+state/diary.reducer';
-import { DiariesComponent, DiariesComponentModule } from './diaries/diaries.component';
+import { DiariesComponent } from './diaries/diaries.component';
 
-@NgModule({
-  imports: [
-    RouterModule.forChild([
-      {
-        path: '',
-        component: DiariesComponent
-      }
-    ]),
-    DiariesComponentModule,
-    StoreModule.forFeature(diaryFeatureKey, diaryReducer),
-    EffectsModule.forFeature([DiaryEffects])
-  ]
-})
-export class DiaryRoutesModule {}
+export const diaryRoutes: Routes = [
+  {
+    path: '',
+    providers: [
+      importProvidersFrom(
+        StoreModule.forFeature(diaryFeatureKey, diaryReducer),
+        EffectsModule.forFeature([DiaryEffects])
+      )
+    ],
+    component: DiariesComponent
+  }
+];