+++ apps/eternal/src/app/newsletter/newsletter.component.module.ts
@@ -1,9 +1,0 @@
-import { NgModule } from '@angular/core';
-import { ReactiveFormsModule } from '@angular/forms';
-import { NewsletterComponent } from './newsletter.component';
-
-@NgModule({
-  imports: [ReactiveFormsModule],
-  declarations: [NewsletterComponent]
-})
-export class NewsletterComponentModule {}