+++ apps/eternal/src/app/newsletter/newsletter.component.spec.ts
@@ -1,70 +1,64 @@
-import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { NewsletterComponent } from './newsletter.component';
+import { ComponentFixture, ComponentFixtureAutoDetect, TestBed } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
-import { createSpyFromClass, Spy } from 'jest-auto-spies';
-import { HttpClient } from '@angular/common/http';
-import { NewsletterService } from './newsletter.service';
 import { By } from '@angular/platform-browser';
-import { asyncScheduler, scheduled } from 'rxjs';
+import { describe, expect, beforeEach } from '@jest/globals';
+import { NewsletterComponent } from './newsletter.component';
 
-describe('Newsletter Component', () => {
-  let fixture: ComponentFixture<NewsletterComponent>;
+describe('NewsletterComponent', () => {
   let component: NewsletterComponent;
-  const getButton = () =>
-    fixture.debugElement.query(By.css('[data-testid=btn-subscribe]'))
-      .nativeElement as HTMLButtonElement;
-  const getInput = () =>
-    fixture.debugElement.query(By.css('[data-testid=inp-email]')).nativeElement as HTMLInputElement;
-  const getInputName = () =>
-    fixture.debugElement.query(By.css('[data-testid=inp-name]')).nativeElement as HTMLInputElement;
-  const getMessage = () =>
-    fixture.debugElement.query(By.css('[data-testid=p-message]'))
-      .nativeElement as HTMLParagraphElement;
-  let newsletterServiceSpy: Spy<NewsletterService>;
+  let fixture: ComponentFixture<NewsletterComponent>;
 
   beforeEach(() => {
-    newsletterServiceSpy = createSpyFromClass(NewsletterService);
-    const httpClient = createSpyFromClass(HttpClient);
-    fixture = TestBed.configureTestingModule({
-      declarations: [NewsletterComponent],
-      imports: [ReactiveFormsModule],
-      providers: [{ provide: NewsletterService, useValue: newsletterServiceSpy }]
-    }).createComponent(NewsletterComponent);
-    fixture.detectChanges();
+    TestBed.configureTestingModule({
+      imports: [NewsletterComponent],
+      providers: [{ provide: ComponentFixtureAutoDetect, useValue: true }]
+    });
+    fixture = TestBed.createComponent(NewsletterComponent);
     component = fixture.componentInstance;
   });
 
-  it('should instantiate', () => {
-    expect(component).toBeInstanceOf(NewsletterComponent);
+  it('should create', () => {
+    expect(component).toBeTruthy();
   });
 
-  it('should test invalid submit', () => {
+  it('should subscribe without DOM interaction', () => {
+    component.formGroup.setValue({ email: 'user@host.com' });
+    component.handleSubmit();
+    expect(component.message).toBe('Thank you for your subscription');
+  });
+
+  it('should not subscribe without DOM interaction', () => {
     component.handleSubmit();
     expect(component.message).toBe('Please provide an email');
   });
 
-  it('should test invalid submit via DOM', fakeAsync(() => {
-    const button = fixture.debugElement.query(By.css('[data-testid=btn-subscribe]'))
-      .nativeElement as HTMLButtonElement;
-    // fixture.autoDetectChanges(true);
-
-    button.click();
-    tick();
-    fixture.detectChanges();
-    const p = fixture.debugElement.query(By.css('[data-testid=p-message]'))
+  it('should not subscribe with DOM interaction', () => {
+    fixture.debugElement.query(By.css('button')).nativeElement.click();
+    const messageBox = fixture.debugElement.query(By.css('p'))
       .nativeElement as HTMLParagraphElement;
-    expect(p.textContent).toBe('Please provide an email');
-  }));
+    expect(messageBox.textContent).toBe('Please provide an email');
+  });
 
-  it('should subscribe successfully via DOM', fakeAsync(() => {
-    newsletterServiceSpy.send.mockReturnValue(scheduled([true], asyncScheduler));
-    const input = getInput();
+  it('should subscribe with DOM interaction', () => {
+    const input = fixture.debugElement.query(By.css('input')).nativeElement as HTMLInputElement;
     input.value = 'user@host.com';
     input.dispatchEvent(new CustomEvent('input'));
 
-    getButton().click();
-    tick();
+    fixture.debugElement.query(By.css('button')).nativeElement.click();
+    const messageBox = fixture.debugElement.query(By.css('p'))
+      .nativeElement as HTMLParagraphElement;
+    expect(messageBox.textContent).toBe('Thank you for your subscription');
+  });
+
+  it('should subscribe in mixed mode', () => {
+    const input = fixture.debugElement.query(By.css('input')).nativeElement as HTMLInputElement;
+    input.value = 'user@host.com';
+    input.dispatchEvent(new CustomEvent('input'));
+
+    component.handleSubmit();
     fixture.detectChanges();
-    expect(getMessage().textContent).toBe('Thank you for your subscription');
-  }));
+    const messageBox = fixture.debugElement.query(By.css('p'))
+      .nativeElement as HTMLParagraphElement;
+    expect(messageBox.textContent).toBe('Thank you for your subscription');
+  });
 });