+++ apps/eternal/src/app/shared/+state/shared.reducer.ts
@@ -1,36 +1,18 @@
 import { createFeature, createReducer, on } from '@ngrx/store';
 import { sharedActions } from './shared.actions';
 
-export interface User {
-  id: number;
-  email: string;
-  firstname: string;
-  name: string;
-  anonymous: boolean;
-}
-
 export interface SharedReducer {
   activeHttpRequest: boolean;
-  loaded: boolean;
-  user: User | null;
 }
 
 const initialState: SharedReducer = {
-  activeHttpRequest: false,
-  loaded: false,
-  user: null
+  activeHttpRequest: false
 };
 
 export const sharedFeature = createFeature({
   name: 'shared',
   reducer: createReducer<SharedReducer>(
     initialState,
-    on(sharedActions.loadUserSuccess, sharedActions.signInUserSuccess, (state, { user }) => ({
-      ...state,
-      user,
-      loaded: true
-    })),
-    on(sharedActions.signOutUserSuccess, (state, { user }) => ({ ...state, user })),
     on(sharedActions.httpRequestStarted, (state) => ({
       ...state,
       activeHttpRequest: true