+++ apps/eternal/src/app/shared/address-lookuper.service.ts
@@ -1,11 +1,12 @@
-import {parseAddress} from "./parse-address";
+import { parseAddress } from './parse-address';
+import { Injectable } from '@angular/core';
 
+@Injectable({ providedIn: 'root' })
 export class AddressLookuper {
-  addresses :string[];
+  addresses: string[] = [];
   constructor(addressesSupplier: () => string[]) {
     this.addresses = addressesSupplier();
   }
-
   lookup(query: string): boolean {
     parseAddress(query);
     return this.addresses.some((address) => address.startsWith(query));