+++ apps/eternal/src/app/shared/loading.service.ts
@@ -1,21 +1,21 @@
-import { Injectable } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { sharedActions } from './+state/shared.actions';
 import { fromShared } from './+state/shared.selectors';
 
 @Injectable({ providedIn: 'root' })
 export class LoadingService {
-  constructor(private store: Store) {}
+  #store = inject(Store);
 
   get isLoading$() {
-    return this.store.select(fromShared.selectActiveHttpRequest);
+    return this.#store.select(fromShared.selectActiveHttpRequest);
   }
 
   loading() {
-    this.store.dispatch(sharedActions.httpRequestStarted());
+    this.#store.dispatch(sharedActions.httpRequestStarted());
   }
 
   loaded() {
-    this.store.dispatch(sharedActions.httpRequestEnded());
+    this.#store.dispatch(sharedActions.httpRequestEnded());
   }
 }